chore: use release yaml files to install the operator
Currently, to install the operator using the plugin we were calling home
(github repo) to retrieve the operator manifest and then apply it, this
PR aims to just load the released manifests and use them internally
to deploy the operator.

For now, it only replace the way to get the manifest, but in the future
we should also allow to modify a couple of things in the manifest.

Closes #2123 #1988

Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>