feat: snapshot recovery from `VolumeSnapshot` (#1960)
This patch allows recovering a CloudNativePG cluster from a set of
CSI VolumeSnapshot via the `.spec.bootstrap.recovery.volumeSnapshot` stanza.

This type of recovery doesn't need a job to be created to initialize the
PVC, since it will already contain the required data - which we assume to
be created by CloudNativePG.

The patch also introduces the `kubectl cnpg snapshot` command which will:

1. select a replica
2. fence it, and wait for PostgreSQL to be shut down
3. create a VolumeSnapshot resource for every PVC used by the replica Pod
4. unfence the replica.

The created set of VolumeSnapshots can be later used to bootstrap a new
CloudNativePG cluster using the `.spec.bootstrap.recovery.dataSource`
stanza.

Closes: #1968

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@gmail.com>
Signed-off-by: Philippe Scorsolini <p.scorsolini@gmail.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Jonathan Gonzalez <jonathan.gonzalez@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>

Co-authored-by: Philippe Scorsolini <p.scorsolini@gmail.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Jonathan Gonzalez <jonathan.gonzalez@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Signed-off-by: netchb <netchb>