feat: improve logging when deleting Pods
The reason to delete a PostgreSQL Pod was registered into the status
stanza of the CR and in an event.

This patch adds a log message which is helpful when diagnosing
Clusters that have been deleted. It also fixes a minor issue: the
operator was adding a event declaring a restart to "apply some
PostgreSQL configuration change" even if the restart was triggered via
the annotation.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@gmail.com>