feat: improve logging when deleting Pods (#2136)
The reason to delete a PostgreSQL Pod was registered into the status
stanza of the CR and in an event.

This patch adds a log message which is helpful when diagnosing
Clusters that have been deleted. It also fixes a minor issue: the
the operator was adding an event declaring a restart to "apply some
PostgreSQL configuration change" even if the restart was triggered via
the annotation.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@gmail.com>
Signed-off-by: Armando Ruocco <armando.ruocco.dev@gmail.com>
Co-authored-by: Armando Ruocco <armando.ruocco.dev@gmail.com>
(cherry picked from commit 48d17af061b8e5aba2d6ffdb12e1b08040b54154)