feat: add schemaOnly to the import section (#2234)
The import feature was previously importing everything, schema and data,
with this new feature we will be able to select only schema or schema and data.

The way is doing this it's using the `section` feature from `pg_dump` and `pg_restore`
commands, we select which section we want to dump/restore, for a full dump and restore
we use `pre-data`, `data` and `post-data` sections, while to import only the schema
the section `pre-data` and `post-data` is used.

Closes #2228

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
(cherry picked from commit ad0fdb27a1f1bcd020256de1f7c5278751b007b7)