fix: early reconcile of Pod metadata (#2242)
Pod labels are fundamental for the functioning of the Cluster
services. This patch ensures that the labels are updated even if the
reconciliation fails at some point.

This prevents failures in following the new primary when an instance
promotion happens on a Cluster having other issues.

Closes #2241

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
(cherry picked from commit 2baf9faf0a51c25782f99ec790917d0a42ae0591)