test: invoke ctrl.SetLogger during TestingEnvironment creation (#2300)
After upgrading to controller-runtime 0.15.0  when the SetLogger() function was not called
we were having a lot of fake error messages saying that this was a bug, fulfilling the screen
with it.

Now, we call the SetLogger() in our internal wrapper of the controller-runtime logs.

Closes #2281

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Co-authored-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
(cherry picked from commit ab1f938417906c86ab8de4815069e1ae4f299753)