fix: improve password expiry management (#2334)
Add special handling for the `VALID UNTIL` role property.

In the case of an existing role to be set not to expire,
the operator will set `VALID UNTIL` to `infinity`
if `VALID UNTIL` was not set to `NULL` in the database.

This is needed because Postgres cannot restore a NULL value
in the `VALID UNTIL` field once you change it.

Closes #2354

Signed-off-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
(cherry picked from commit ba2545d78a57d6369527c2aee6f308924b2ac023)