chore: align structs to reduce space usage (#2458)
Some of the structs were not aligned and therefore using some more bytes
and cachelines than were actually needed.

Make Pod variable inside PostgresqlStatus a pointer to reduce the size
of the struct in 1056 bytes reducing the amount of space that we move
around every time we query the status of the PostgreSQL cluster.

Closes #2457

---------

Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
Signed-off-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Signed-off-by: John Long <john.long@enterprisedb.com>
Co-authored-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Co-authored-by: John Long <john.long@enterprisedb.com>