fix: backup reconciliation logic for running backups (#2591)
Before this patch, the Backup resource reconciliation cycle was
exiting at the start when the backup Phase was not pending, making it
impossible to reconcile interrupted backups.

With this patch, the reconciliation cycle stops only if the Backup is
in a final Phase (Completed or Failed).

Closes #2467

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Co-authored-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
(cherry picked from commit 43e2268bf3fee0a658e4a571ec1aa4e56bca7ea8)