chore: move api documentation generation from k8s-api-docgen to genref (#2606)
Use genref to create the documentation instead of k8s-api-docgen. prometheus-operator abandoned the original version of it for gen-crd-api-reference-docs, and that project now suggests to use alternatives due to lack of time for maintenance.

k8s-api-docgen is based on omitempty directives, while genref uses +optional, so this commit includes a review of the API to align the two. It also specifies as mandatory some field that were actually required but not defined as such.

Closes #2350

Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>