chore: add coredumpFilter annotation
Since it's useful to reduce the amount of memory used by a coredump
when it's created we added this annotation to set the value for
/proc/self/coredump_filter. The default value it's now set to 0x31
since it's look like the most used value.

The annotation can be set at cluster level and then will be passed
in all the process that run any PostgreSQL command.

Closes #2696

Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>