fix(monitoring): run `pg_replication_slots` metric on primary only (#2723)
The `pg_replication_slots` metric that is shipped by default
with the operator is consistently failing on the replicas due
to calls to WAL functions that are not available in the recovery
phase. The metric is now run only on the primary.

Closes #2377
Closes #2552

Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
(cherry picked from commit 23db603f07cbb35ec60d7bd6c140a4f79640f9a0)