fix(monitoring): improve accuracy of replica lag
This patch overrides the previous patch which only covered the
documentation samples. From a deeper analysis, we noticed that
some examples had already been introduced in the default monitoring
config map, thus have been removed from the docs.

The default configmap now fixes the `pg_replication` metric to use the
`pg_last_wal_receive_lsn()` and `pg_last_wal_replay_lsn()` functions to better
estimate the lag of a replica, and avoid the increase of the lag on inactive
systems.

Closes #1814

Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>