chore: add `cnpg.io/coredumpFilter` annotation (#2733)
Introduce the `cnpg.io/coredumpFilter` annotation to control
the content of a core dump generated in the unlikely (but possible)
event of a PostgreSQL crash.

The annotation follows the convention defined by the Linux Kernel
through `/core/self/coredump_filter` - for more information see:

    https://docs.kernel.org/filesystems/proc.html#proc-pid-coredump-filter-core-dump-filtering-settings

By default the filter is set to `0x31` in order to exclude shared memory
segments from the dump. This setting only takes effect during Pod
startup.

The annotation can be set at cluster level. A change of the annotation
doesn't trigger an automated rollout of the instances.

Closes #2696

Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
Signed-off-by: Tao Li <tao.li@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Tao Li <tao.li@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>