feat: change default value of `startDelay` to 3600 (#2847)
IMPORTANT: The default value of `startDelay` has now been
changed to 3600 seconds, equivalent of 1 hour.

Previously, it was set to 30 seconds.

This patch also replaces the livenessProbe's initial delay with a
more proper Kubernetes startup probe to deal with the start of
a Postgres server. Both the startup probe and the higher default
time should improve the self-healing capabilities of larger Postgres
deployments where the default value of 30 seconds was not
sufficient for example to complete a crash recovery operation,
causing infinite restarts.

Closes #2843

Signed-off-by: Tao Li <tao.li@enterprisedb.com>
Signed-off-by: YanniHu1996 <yantian.hu@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: YanniHu1996 <yantian.hu@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>