feat: change default value of `stopDelay` to 1800 (#2848)
IMPORTANT: The default value of `stopDelay` has now been
changed to 1800 seconds, equivalent of half an hour.

Previously, it was set to 30 seconds.

This patch also introduces a new parameter, called `smartStopDelay`,
to control the window of time reserved for the smart shutdown of
Postgres to complete. The general formula to compute the overall
timeout to stop Postgres is `max(stopDelay -  smartStopDelay, 30)`.

Closes #2844

Signed-off-by: YanniHu1996 <yantian.hu@enterprisedb.com>
Signed-off-by: Tao Li <tao.li@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Tao Li <tao.li@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Signed-off-by: Tao Li <tao.li@enterprisedb.com>