tests: ensure client certificate is always being created (#2870)
The connectivity test were failing due to a dependency of a previous test
that should create the proper secret with the required certificate. Now,
we create the certificate every time in the test to keep the tests
independent.

Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
Signed-off-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Co-authored-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
(cherry picked from commit 7125b96512a52726321de52ad5f70741ae52d39e)