feat: reconcile cluster services
Starting from this patch, CloudNativePG reconciles the Services every
time the selectors change or when the labels or the annotations managed
by the operator needs to be changed.

From now on, we can fix the services created by old versions to the new
`cnpg.io/cluster` label available on Pods.

Closes #2917

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>