fix: disable slots in cascading replication (#2960)
In case replication slots for HA were enabled, the designated primary
in a replica cluster was requesting to use the slot with their origin
(the source in the external cluster), resulting in an failure (as the
replication slot is missing on the other side).

The operator should not use the same setting to control that.
This is indeed a different feature, which should be controlled by
a different option.

Closes #2959

Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
(cherry picked from commit 1460a461f2fae10b58be29b4672388a7921c745d)