fix: prevent stuck cluster when primary database is down but pod is up (#2966)
Fixes a condition that could result in the primary database being down
while the corresponding pod is still up.

When this happens, the reconciliation of replication slots and managed
roles will error out, preventing the instance manager from shutting down
the primary and preventing switch-overs and fail-overs from finishing
correctly.

Signed-off-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>