fix: defaults to `app` owner and database during recovery (#2957)
This patch makes the defaulting webhook use `app` for the database
owner and the database name when the fields are not already specified
by the user.

This is symmetrical to the `initdb` section and ensures that a cluster
being recovered always has a database name and a database owner.

Closes #2425

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
(cherry picked from commit 0df251b686e1556a250fda32709307d093bac434)