feat: move shutdown logic to instance management (#3039)
This patch moves the shutdown logic from the lifecycle manager to the
instance manager, simplifying its structure.

The `.spec.smartStopDelay` field is used now as the time given to the
PostgreSQL smart shutdown sequence. After that, we'll request PostgreSQL
to shut down using the `fast` strategy within the overall `.spec.stopDelay`
time, which is the `terminationGracePeriodSeconds` of the Pods.

Closes #3032

Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>

Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>