ci: use goreleaser snapshot mode when tags are unavailable (#2247)
By default, GoReleaser requires the git tags in the local repository to run a build.
However, the tags are often unavailable when running from a fork, and the build fails.

With this patch, we pass the `--snapshot` option GoReleaser in such cases,
the build to succeed.

Closes #2179

Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
(cherry picked from commit 6a98534be4cd7bcdc3533fba8db421660277101b)