feat: use extended query protocol for PostgreSQL
This patch makes the operator use two different connection profiles for
PostgreSQL and for PgBouncer. The first one is using the extended query
protocol, the second one the simple one.

The reasoning behind is that for PostgreSQL we'll need to use the more
secure query cancelation method. PgBouncer needs the simple query
protocol against the admin DB, and we use that for metrics.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>