fix: defer WAL archiving during a recovery job (#3251)
Up to now, the recovery job was discarding WAL files by setting
`archive_command` to `cd .` (that is success). We now set it to `false`,
so that we defer the decision about archiving a WAL file to when the
instance is actually started as a regular pod.

Closes #3259

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
(cherry picked from commit 7e46400925637069e536ee6210372267e3d2499f)