chore: allow one more connection to PostgreSQL from the instance manager
When taking an hot snapshot backup, we require an exclusive connection
for the PostgreSQL Physical Backup API.

In that case, the number of available connections for the other instance
manager runnables will be reduced to one. To avoid those other runnables
to be stuck, we allow one more PostgreSQL connection.

Closes: #3275

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>