ci: do not update the "install the latest code" link (#3283)
During the release procedure, we update the link to install the latest
snapshot of the operator together with the other. That link should
always point to the newest snapshot from the main branch.

This patch stops changing the "install the latest code" link in the release script.

Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
(cherry picked from commit a45710691837e2a622a039a453326851fe4db920)