fix: correctly evaluate the request and response for snapshot backups (#3321)
Fixes a bug in the backup snapshot code where an error reading the body would
be handled as an overall error. This could lead to looping behavior stuck on
error.

With the new implementation, the caller determines if the returned body error
is relevant to the backup that is trying to run.

The patch also ensures that requests for outdated backups aren't processed.

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Co-authored-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
(cherry picked from commit 4bca50b5fbd396d94f5c3384d4766bd65b6b0a08)