feat: declarative tablespaces support (#3410)
Introduce the `tablespaces` stanza in the `Cluster` spec to provide full
lifecycle management of PostgreSQL tablespaces, by operating at two
distinct levels:

- Kubernetes: managing persistent volume claims, identically to how
PGDATA and WAL volumes are handled
- PostgreSQL: managing the `TABLESPACE` global objects within the
PostgreSQL instance and providing an abstraction to the related DDL
commands, such as `CREATE TABLESPACE` and `ALTER TABLESPACE`.

As part of the Kubernetes ecosystem, CloudNativePG's declarative
tablespaces are implemented leveraging Persistent Volume Claims (and
Persistent Volumes). Each tablespace defined in the cluster is housed in
its own persistent volume. CloudNativePG takes care of generating the
PVCs, mounting the required volumes in the instance Pods in normalized
locations, and ensuring replicas are ready to support tablespaces before
activating them in the primary.

Tablespaces can be setup when the cluster is created, or added at a
later time - provided the storage is available when requested. By
default, they are owned by the application user, but the ownership can
be changed.

Currently, tablespaces cannot be removed, but this limitation will be
addressed in a future minor/patch version of CloudNativePG.

Partially resolves #11

Signed-off-by: Tao Li <tao.li@enterprisedb.com>
Signed-off-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
Co-authored-by: Tao Li <tao.li@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>