fix: avoid crashing with PodMonitor CRD but no monitoring section (#3514)
This patch fixes an issue happening when a Kubernetes cluster has the PodMonitor
CRD installed (suggesting the availability of the Prometheus operator), but the
"monitor" section is not configured. The operator will panic when this happens
in the Cluster or Pooler resources.

This patch fixes the issue, preventing the operator from panicking.

Closes: #3515

Signed-off-by: YanniHu1996 <yantian.hu@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>