fix: verify candidate has active wal_receiver before targeting as primary (#3456)
The current implementation did not consider whether a target primary for a
switchover had an active wal_receiver session. This could lead to potential
situations where a Pod could be selected as the target primary even if it is in
the process of running wal-restore to catch up.

This could have been avoided if the replica with the active wal_receiver
had been selected, and this is what this patch is doing.

Signed-off-by: Jakob Schultz-Falk <sjulsbe@gmail.com>
(cherry picked from commit 4772a70b5c8beb1e706bd4eb624ce60df95bec15)