feat: image catalog
Add the CRDs ImageCatalog and ClusterImageCatalog. This will allow
users to define an image to use for each major version of PostgreSQL,
in namespace-scoped or cluster-scoped resources.
Changing the image in the catalog will trigger updates for the clusters
using it.
The major version is included in the catalog, and can be be
leveraged to avoid the image naming restriction `imageName` has.

Closes #12

Signed-off-by: Francesco Canovai <francesco.canovai@enterprisedb.com>