feat: add `pgadmin4` command to `cnpg` plugin (#3701)
Introduce the `pgadmin4` command in the `cnpg` plugin for `kubectl`.
This provides a simple way to connect to a specific database `Cluster`
and browse its content in a local environment, such as `kind`,
for demonstration and evaluation purposes.

This patch implements two modes for deploying pgAdmin4:

- `server` mode: authentication is required using a randomly generated
   password, and users must manually specify the database to connect to
- `desktop` mode: initiates a pgAdmin web interface without requiring
authentication; it automatically connects to the `app` database as the
  `app` user, making it ideal for quick demos

Closes: #3696

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
(cherry picked from commit 02e420d7af0004f3a23bd3cad329135df9fba0f0)