chore: improve OLM environment detection
As far as now, we were relying on the detection of the SecurityContextConstraint
to know that we were running on an OLM environment but this is not a secure way
since an OLM environment may or may not have a SCC inside, that's why we should
look at the `operators.coreos.com/v1` resources and check that  it exists, if this
is the case, we're in an OLM environment.

This helps us to improve the plugin report and properly retrieve more useful
information.

Closes #3764

Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>