chore: make clear which Dockerfile use every build process
Until now, we were supposing that we were using Dockerfile by default
and just specify when we wanted to use Dockerfile-ubi8, now, we're
explicity on which file we use, this is for in the future, avoid making
mistakes of which file we're using to build the container images.

And since we're using more than one Dockerfile to build image, we also
added a new stop to build that image and run the Dockle scan process
on that one.

Closes #3811

Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>