fix: reduce the time required for the tablespaces test (#3911)
With this patch, we make sure that the WAL files are all archived after
we take the backup; this will reduce the amount of time required by
the restore process when the backup is taken from standby, as it
will not wait for the WAL files to be archived on the primary.

The time is reduced from ~9m to ~4m

This patch is a quick fix to improve the time of our tests, but we need
to refactor it in a broader initiative about backup testing E2E test
improvements.

Closes #3910

Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
(cherry picked from commit 167caebd6e25c2bef3453ca128dc21d934040609)