fix: add --force-conflicts options to kubectl apply
While doing the upgrade process if we don't add the `--force-conflicts` along
side with the `--force-conflicts` we can't apply the new manifest meaning
that the upgrade test fails. Now we added this `--force-conflicts` option
that we already documented in the upgrade process.

More information here https://kubernetes.io/docs/reference/using-api/server-side-apply/#conflicts

Closes #4000

Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>