fix: Make sure PVCs have correct value of the label `instanceRole` and 'Role' (#3930)
This patch makes sure that the PVCs labels are always synchronized with the
labels on the Pods. This is important when all the pods are deleted and the operator
needs to decide which Pod recreate first.

Closes: #3810

Signed-off-by: YanniHu1996 <yantian.hu@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
(cherry picked from commit 2ed27a98efd6f6c6e26e1fa5721b3d559087a375)