fix: prevent endless reconciliation when unsetting connection limit in roles (#3832)
In the managed roles reconciler, the connection limit was only
explicitly set if it was not default (-1). This does not work as
intended if the managed role tries to unset a connection limit
that was not previously -1.

In this case the instance manager gets into indefinite reconciliation:

- it recognizes the value in the DB is not the same as the manifest
- since the manifest has the default connection limit, it makes no
  update

This patch always explicitly sets the connection limit when updating
a role.

Closes #3831

Signed-off-by: Jaime Silvela <jaime.silvela@enterprisedb.com>
(cherry picked from commit 4d8516b7b29e62e60c581bebba66841079dbb0b3)