fix: ensure pods with no ownership are deleted during cluster restore
At present, during the cluster restore from orphan PVCs , it is not possible for the operator to bring back a cluster to a functional state if any orphan pods are present.

This is because the existing pods will linger indefenetly and won't be be recreated with the correct data. To address this issue, this patch ensures that no orphan pods are present during the cluster restore from PVCs by deleting them.

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>