refactor: improve metadata fencing code organization
Currently, the metadata fencing code is challenging for inexperienced developers to use, given that it is fragmented across the codebase without a clear pattern.

This patch aggregates the logic and adds a fluent builder to execute the fencing metadata logic.

It also improves how we handle certain operations like fencing or unfecing an already present or deleted instance. In those cases, the executor does not return an error anymore

Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>