fix: align enableSuperuserAccess default
Within the `GetEnableSuperuserAccess` function, the codebase defaulted
to `true` while the `cluster.spec.enableSuperuserAccess` in the CRD was
defaulted to `false`.

This discrepancy does not lead to any runtime issues because when the
operator retrieves Custom Resources from the API server, the structure
has already been populated with the correct values.

However, the problem can still surface during unit tests when utilizing
a mock Kubernetes client.

Closes: #4211

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>