feat: allow configuring wal_log_hints
This patch enables the ability to set wal_log_hints to "off" for
development clusters with a single instance.

Previously, the operator set wal_log_hints to "on" for every cluster due
to pg_rewind’s requirement for reattaching an old primary instance.

However, since this requirement doesn't apply to single-instance
clusters, we've loosened the restriction for these cases.

Closes: #4041

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>