feat(pod-monitor-interval): implemented PodMonitor specs and fix a regression
This commit fix a bug introduced by the Plugin API which caused the reconcile failure for object not recognized by the operator, such as PodMonitor. The lifecycle hook now stop if the resource is not known but returns the original object so the reconcile can continue.

Furthermore, this commit introduce new specs for PodMonitor functionality with the plan to implement 1:1 functionality as PodMonitor spec by also guard-railing the creation. This has been done by completely rewrite the PodMonitor logic implementing a new kind of Controller that will implement all the current and future logic.

This commit also propose the deprecation of the original PodMonitor field in favour of a more standardized sub-spec which can be improved more easily in the future; finally e2e test for PodMonitor are implemented to intercept eventual regression more easily.

Signed-off-by: Nicolò Ciraci <nicolo.ciraci@elmec.it>