fix(pod-monitor): Fix regression for PodMonitor
This commit fix a regression introduced by the new Plugin API which caused the PodMonitor - and probably other CRD - to not be created. Some E2E test have been introduced to detect issues with PodMonitors.

Furthermore some commodities have been added and general code tweaks.

Kindly note that this commits is a cherry picked version of #4268, which contains a broader review of PodMonitor specs.

Signed-off-by: Nicolò Ciraci <ciraci.nicolo@gmail.com>