fix: empty LSN handling in instance sorting
`receivedLSN` is empty for instances that never attached to streaming
replication.

The instance sorting code considers this an error and the `Less`
implementation for LSNs is not antisymmetric, resulting in a non-stable
sorting.

This could result in the operator preferring a replica which never
attached to streaming replication over a replica which is more advanced
but attached to streaming replication.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>