fix: empty LSN handling in instance sorting (#4283)
`receivedLSN` is empty for instances that have never been attached to
streaming replication.

The instance sorting code considers this an error, and the `Less`
implementation for LSNs is not antisymmetric, resulting in a non-stable
sorting.

This could result in the operator preferring a replica that is never
attached to streaming replication over a replica that is more advanced
but attached to streaming replication.

Closes: #4284

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
(cherry picked from commit 3e8fea06a0da792b12e6a09b585d7cf5a8c5dcfb)