fix: wait for elected replica to be in streaming mode before a switchover (#4288)
Before switching over to a replica to reconfigure the primary,
we now wait for the elected replica to be connected in streaming mode.

We do that to ensure that the replica receives all the data written
on the primary when it shuts down.

Closes: #4285

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
(cherry picked from commit 2edfa8873863672800ecd193e026c06472279597)