feat: automatically fence the primary when no disk space is left
PostgreSQL will shutdown cleanly when there is no enough disk space to
store WAL files.

The operator was not recognizing this condition and, since the primary
failed, was performing a failover to the most advanced replica. This
action will not fix the underlying issue.

Only a manual disk resize, initiated by the user, can ultimately lead
to a fully working PostgreSQL cluster.

This patch makes the instance manager recognize this condition, and
report it back to the operator. Upon detecting it, the operator will
fence the primary instance and set a phase describing the situation.
Since the primary instance is fenced, no failovers will be done.

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enteprisedb.com>