fix: instance manager stuck if PostgreSQL doesn't start
If a primary PostgreSQL instance fails to start after being stopped, the
instance manager is indefinitely waiting for it to start running the
configuration queries.

This patch creates a separate context for each postmaster process and
runs the configuration queries in a parallel goroutine, canceling it as
soon as the postmaster finishes.

Fixes: #4433

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enteprisedb.com>