fix: instance manager stuck if PostgreSQL doesn't start (#4434)
If a primary PostgreSQL instance fails to start after being stopped, the
instance manager is indefinitely waiting for it to start running the
configuration queries.

This patch creates a separate context for each postmaster process and
runs the configuration queries in a parallel goroutine, canceling it as
soon as the postmaster finishes.

Fixes: #4433

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enteprisedb.com>
Signed-off-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Signed-off-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enteprisedb.com>
Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
(cherry picked from commit 473e5e72356644c3eff6711a4562a027a9f1cf5d)