fix: avoid reload PG when the parameters didn't change
The instance manages stores the PostgreSQL configuration inside a
dictionary, mapping GUC names to their actual value.

When writing the configuration file, the code is iterating over a map
and since the iteration order is not defined, was re-writing the
configuration file even when the parameters didn't really change.

This made the instance manager SIGHUP the postmaster everytime, even if
this is not really needed.

This patch fixes the issue by writing che configuration file in
lexicographic order.

Closes: #4529

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>