fix: increase the timeout for the failover/switchover tests (#4532)
The walreceiver of a standby that wasn't promoted may try to reconnect
before the rw service endpoints are updated. In this case, the walreceiver
can be stuck for waiting for the connection to be established for a time that
depends on the tcp_syn_retries sysctl. Since by default
net.ipv4.tcp_syn_retries=6, PostgreSQL can wait 2^7-1=127 seconds before
restarting the walreceiver.

Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
Signed-off-by: Jonathan Gonzalez V. <jonathan.abdiel@gmail.com>
Co-authored-by: Francesco Canovai <francesco.canovai@enterprisedb.com>